# A slightly faster interpose.