@eriknw 2017-07-04
Looks good to me.  Merging.  Thanks, @groutr!